requirements.txt
requests
beautifulsoup4
numpy
aiohttp
pysmb
validators
python-dotenv
google-api-python-client
pillow
duckduckgo-search
torch
torchvision
realesrgan